Use ViewComponent instead
