+++ src/OrchardCore.Modules/OrchardCore.Localization/ViewModels/CulturePickerViewModel.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.Globalization;
+
+namespace OrchardCore.Localization.ViewModels
+{
+    public class CulturePickerViewModel
+    {
+        public string SelectedCulture { get; set; }
+
+        public IList<CultureInfo> SupportedUICultures { get; set; }
+    }
+}
